refactor: Update fetch_rss_feed to fetch a specified number of entries
The fetch_rss_feed function in tools.py has been updated to accept an optional parameter, num_entries, which allows the user to specify the number of entries to fetch from the RSS feed. The default value is set to 5. This change provides more flexibility in fetching the desired number of latest entries from the RSS feed.